+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @under-control/core
 
+## 1.7.23
+
+### Patch Changes
+
+- add ref to controlled
+
+## 1.7.22
+
+### Patch Changes
+
+- Add forwardRef to controlled
+
 ## 1.7.21
 
 ### Patch Changes